小程序消息模板

+++ database/migrations/2019_04_23_135825_min_app_template_message.php
@@ -20,7 +20,7 @@ class MinAppTemplateMessage extends Migration
                 $table->string('title');
                 $table->string('template_id', 45);
                 $table->text('data', 65535)->nullable();
-                $table->tinyInteger('is_default')->nullable();
+                $table->integer('is_default')->nullable();
                 $table->integer('is_open')->default(0);
                 $table->integer('created_at')->nullable();
                 $table->integer('updated_at')->nullable();
